iio: magnetometer: add mmc34160 magnetometer driver
This patch adds support for the MMC34160 3-axis magnetometer driver. The
MMC31460 magnetometer driver uses the same register map as MMC35240 with
different specifications for sensitivity.

According to Memsic specification, for the MMC31460 sensor, there is no
need to apply compensation to the read values. Also, the MMC34160 sensor
does not use the same formula as MMC35240 for transforming raw data into
mgauss, and the current formula is based on the input driver (mmc3416x)
provided by Memsic.

Signed-off-by: Teodora Baluta <teodora.baluta@intel.com>